third_party/WebKit/LayoutTests/fast/forms/file/file-input-webkitdirectory-key-enter-prevent-keypress-expected.txt

CONSOLE MESSAGE: line 5: FileChooser: opened; mode=kUploadFolder
CONSOLE MESSAGE: line 5: FileChooser: mock-file-chooser.js doesn't support directory selection yet.
Ensure pressing Enter key with focus on an INPUT element with TYPE=FILE and WEBKITDIRECTORY still launches a file chooser if preventDefault() is called on the 'keypress' event. Automated test passes if 'FileChooser: opened' was logged.

keypress prevented
